+++ modules/tilemap.py
@@ -2,13 +2,15 @@ import pygame
 NEIGHBOR_OFFSET = [(-1,0),(-1,-1),(0,-1),(1,-1),(1,0),(0,0),(-1,1),(0,1),(1,1),]
 COLLIDABLE_TILE_TYPES = {"grass", "stone"}
 from modules.utils import sum_vectors, k_vector
+import json
 
 
-def pos_in_pixels(pos_normal, tile_size=16):
-    return tuple(map(lambda x: x*tile_size, pos_normal))
+def pos_in_pixels(posicao_tilemap, tile_size=16):
+    """retorna a posicao em pixels a partir da posicao no tilemap"""
+    return tuple(map(lambda x: x*tile_size, posicao_tilemap))
 
 class Tilemap:
-    def __init__(self, game, tile_size=16):
+    def __init__(self, game, map_filename="", tile_size=16):
         self.game = game
         self.tile_size = tile_size
         #temos dois sistemas: um que é em uma grid (grade)
@@ -17,10 +19,27 @@ class Tilemap:
         #esses offgrids sao decoracoes, background, basicamente
         self.offgrid_tiles = []
 
-        for i in range(10):
-            self.tilemap[str(3 + i) + ";10"] = {'type': 'grass', 'variant': 1, 'pos': (3 + i, 10)}
-            self.tilemap["10;" + str(5 + i)] = {'type': 'stone', 'variant': 1, 'pos': (10, 5 + i)}            
-    
+
+        # preenchendo o mapa
+        if map_filename:
+            #abrindo o arquivo e carregando as informacoes
+            with open(map_filename, "r") as map_F:
+                raiz = json.loads(map_F.read())
+                tilemap_arquivo = raiz['tilemap']
+                tile_size = raiz['tile_size']
+                offgrid_tilemap = raiz['offgrid']
+
+            #guardando os tiles do arquivo no nosso tilemap
+            #observe que cada tile esta armazenado num dicionario cuja chave é a coordenada e os valore sao {'type': 'stone', 'variant': 1, 'pos': (10, 5 + i)}  
+            for coordenada in tilemap_arquivo:
+                self.tilemap[coordenada] = tilemap_arquivo[coordenada]      
+        else:
+            #preenchendo da maneira padrao
+            for i in range(10):
+                self.tilemap[str(3 + i) + ";10"] = {'type': 'grass', 'variant': 1, 'pos': (3 + i, 10)}
+                self.tilemap["10;" + str(5 + i)] = {'type': 'stone', 'variant': 1, 'pos': (10, 5 + i)}       
+                
+
     def neighbor_tiles(self, pos):
         """Retorna uma lista com no máximo todos os 9 blocos ao redor do personagem, se existirem"""
         # a posicao esta em pixel, passemos-na pra posicao em tiles
@@ -58,9 +77,12 @@ class Tilemap:
             tile = self.tilemap[loc]
             tile_x, tile_y = tile['pos'] 
             if tile_x >= start_screen_x  and tile_x <= end_screen_x:
+                #pegar a imagem dos assets daquele tipo (grass):  tile['type]
+                #pegar a imagem de numero = variant (1) desta lista
+                # e desenhá-la na tela
                 surface.blit(self.game.assets[tile['type']][tile['variant']],
                             sum_vectors(k_vector(-1, offset), 
-                            tuple(map(lambda x: x*self.tile_size, tile['pos'])))
-        )
+                            pos_in_pixels(tile['pos'])
+                            ))